requirements.txt
torch
torchvision
torchaudio
supervision
opencv-python
numpy
Pillow
tqdm
ultralytics
transformers
inference-sdk
rfdetr
